backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
boto3==1.24.71
pytest==7.0.0
pytest-flask==1.2.0
pytest-cov==4.1.0
python-dotenv==0.19.2
PyJWT==2.3.0
requests==2.27.1
Werkzeug==2.3.7
cryptography==37.0.4
gunicorn==20.1.0